fix init error

+++ src/ubsocket/csrc/core/umq/umq_setting.h
@@ -36,8 +36,8 @@ public:
     static uint16_t UMQ_FC_DEFAULT_CREDIT;
     static uint16_t UMQ_FC_MAX_CREDIT;
     static uint16_t UMQ_FC_MIN_CREDIT;
-    static uint16_t UMQ_MEM_POOL_INIT_SIZE_MB;
-    static uint16_t UMQ_MEM_POOL_MAX_SIZE_MB;
+    static uint64_t UMQ_MEM_POOL_INIT_SIZE_MB;
+    static uint64_t UMQ_MEM_POOL_MAX_SIZE_MB;
     static uint64_t UMQ_BUF_POOL_DEPTH;
     static uint32_t UMQ_POST_BATCH_MAX;
     static umq_buf_block_size_t IO_BLOCK_TYPE;
